# src/main/java/com/openbanking/balances/model/OBReadBalance1.java
package com.openbanking.balances.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadBalance1
 */
@Getter
@Setter
public class OBReadBalance1 implements Serializable {
	private static final long serialVersionUID = 1L;

}